requirements.txt
matplotlib==3.8.2
pandas==1.4.3
Pillow==10.2.0
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
